fix(client): edit game in progress folder achitecture
