--- packages/client/src/screens/Room/GameInProgress/Answer/index.ts
@@ -1,3 +1,0 @@
-import GameInProgressAnswer from './GameInProgressAnswer';
-
-export default GameInProgressAnswer;
