+++ packages/client/src/screens/Room/GameInProgress/GameInput/GameInputStyle.tsx
@@ -1,0 +1,18 @@
+import { fontSizes, fontFamilies } from '@Src/constant/theme';
+import { StyleSheet, Platform } from 'react-native';
+
+const styles = StyleSheet.create({
+  container: {
+    flexDirection: 'row',
+    padding: 14,
+  },
+  input: {
+    ...(Platform.OS === 'web' && { outlineWidth: 0 }),
+    fontFamily: fontFamilies.text,
+    fontSize: fontSizes.xxl,
+    flexGrow: 1,
+    minWidth: 10,
+  },
+});
+
+export default styles;
