+++ packages/client/src/screens/Room/GameInProgress/GameInput/index.ts
@@ -1,0 +1,3 @@
+import GameInput from './GameInput';
+
+export default GameInput;
