+++ packages/client/src/screens/Room/GameInProgress/Question/QuestionStyle.tsx
