+++ packages/client/src/screens/Room/GameInProgress/RoundEnd/Answer/index.ts
@@ -1,0 +1,3 @@
+import Answer from './Answer';
+
+export default Answer;
